# context/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Enter the access token returned by /api/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "API Support",
      "url": "https://github.com/bruno-frascino/piggy-api"
    },
    "description": "A comprehensive stock portfolio management API built with Express and Prisma",
    "title": "Truffles API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/accounts": {
      "get": {
        "responses": {
          "200": {
            "description": "Trading accounts"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List trading accounts for the authenticated user",
        "tags": [
          "Accounts"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "Main",
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Existing account returned"
          },
          "201": {
            "description": "Account created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Exchange not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create (or return existing) trading account for an exchange",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}": {
      "delete": {
        "description": "Permanently deletes a trading account owned by the authenticated user, along with its derived performance history (portfolio snapshots). Refuses to delete an account that still has positions, or one that is referenced by a generated tax report — close the account instead to preserve that history.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account has positions or is used by a tax report"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete a trading account",
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "description": "Rename a trading account",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "Updated Account Name",
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account name already exists for this user"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a trading account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/close": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account closed"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account still has open positions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Close a trading account while preserving history",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}/reopen": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account reopened"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Reopen a closed trading account",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:4000"
    }
  ],
...
}